refactor: move all type utilities to one file
