+++ src/validators/ArrayValidator.ts
@@ -14,6 +14,7 @@ import type { BaseError } from '../lib/errors/BaseError';
 import { CombinedPropertyError } from '../lib/errors/CombinedPropertyError';
 import { ValidationError } from '../lib/errors/ValidationError';
 import { Result } from '../lib/Result';
+import type { ExpandSmallerTuples, Tuple, UnshiftTuple } from '../lib/util-types';
 import { BaseValidator } from './imports';
 
 export class ArrayValidator<T> extends BaseValidator<T[]> {
@@ -96,19 +97,3 @@ export class ArrayValidator<T> extends BaseValidator<T[]> {
 			: Result.err(new CombinedPropertyError(errors));
 	}
 }
-
-export type UnshiftTuple<T extends [...any[]]> = T extends [T[0], ...infer Tail] ? Tail : never;
-export type ExpandSmallerTuples<T extends [...any[]]> = T extends [T[0], ...infer Tail] ? T | ExpandSmallerTuples<Tail> : [];
-
-// https://github.com/microsoft/TypeScript/issues/26223#issuecomment-755067958
-export type Shift<A extends Array<any>> = ((...args: A) => void) extends (...args: [A[0], ...infer R]) => void ? R : never;
-
-export type GrowExpRev<A extends Array<any>, N extends number, P extends Array<Array<any>>> = A['length'] extends N
-	? A
-	: GrowExpRev<[...A, ...P[0]][N] extends undefined ? [...A, ...P[0]] : A, N, Shift<P>>;
-
-export type GrowExp<A extends Array<any>, N extends number, P extends Array<Array<any>>> = [...A, ...A][N] extends undefined
-	? GrowExp<[...A, ...A], N, [A, ...P]>
-	: GrowExpRev<A, N, P>;
-
-export type Tuple<T, N extends number> = number extends N ? Array<T> : N extends 0 ? [] : N extends 1 ? [T] : GrowExp<[T], N, [[]]>;
